fixes most of the remaining routes
