+++ workspaces/idapp/app/(tabs)/settings/confirm-factory-reset.tsx
@@ -1,0 +1,11 @@
+import { ErrorBoundary } from 'react-error-boundary'
+import { Diagnostics } from '~/views/diagnostics'
+import { ConfirmFactoryReset } from '~/views/settings'
+
+export default function ConfirmFactoryResetRoute() {
+  return (
+    <ErrorBoundary FallbackComponent={Diagnostics}>
+      <ConfirmFactoryReset />
+    </ErrorBoundary>
+  )
+}
