+++ workspaces/idapp/app/(tabs)/settings/diagnostics.tsx
@@ -1,0 +1,5 @@
+import { Diagnostics } from '~/views/diagnostics'
+
+export default function DiagnosticsRoute() {
+  return <Diagnostics />
+}
