+++ workspaces/idapp/src/views/main/index.ts
@@ -1,2 +1,3 @@
 export { Main } from './Main'
 export { AppLoading } from './AppLoading'
+export { ScanIdBox } from './ScanIdBox'
